Add files via upload
This is the code I used to clean and modify the dataset and create the graphs I used for the PDF file. This also answers the questions to the assessment, but it is more difficult to parse through due to the amount of information. The file also works better if you have a program to view Jupyter Notebooks